Rename MeetingService to meetingService
