+++ mellophone/frontend/src/pages/CreateMeeting.tsx
@@ -4,7 +4,7 @@ import { RouteComponentProps } from "@reach/router";
 import Main from "../elements/Main";
 import Route from "../utils/Route";
 import { IMeetingToBeCreated } from "../types";
-import MeetingService from "../network/MeetingService";
+import meetingService from "../network/meetingService";
 import CreateMeetingForm from "../sections/CreateMeetingForm";
 import requireAuthentication from "../utils/requireAuthentication";
 
@@ -14,7 +14,7 @@ function CreateMeeting(props: RouteComponentProps<{ teamId: string }>) {
   if (Number.isNaN(teamId)) return null;
 
   const createMeeting = async (meeting: IMeetingToBeCreated) => {
-    const meetingId = (await MeetingService.createMeeting(meeting, teamId)).id;
+    const meetingId = (await meetingService.createMeeting(meeting, teamId)).id;
     new Route(Route.MEETINGS, meetingId).navigate();
   };
