+++ mellophone/frontend/src/pages/Meeting.tsx
@@ -4,7 +4,7 @@ import { RouteComponentProps } from "@reach/router";
 import { IMeeting } from "../types";
 import Main from "../elements/Main";
 import MeetingDocument from "../sections/MeetingDocument";
-import MeetingService from "../network/MeetingService";
+import meetingService from "../network/meetingService";
 import requireAuthentication from "../utils/requireAuthentication";
 
 function Meeting(props: RouteComponentProps<{ meetingId: string }>) {
@@ -14,7 +14,7 @@ function Meeting(props: RouteComponentProps<{ meetingId: string }>) {
 
   React.useEffect(() => {
     if (!Number.isNaN(meetingId)) {
-      MeetingService.getMeetingById(meetingId)
+      meetingService.getMeetingById(meetingId)
         .then(meeting => setMeeting(meeting))
         .catch(
           error => process.env.NODE_ENV !== "production" && console.error(error)
